requirements.txt
crewai
langchain_google_genai
composio_crewai
discord.py
python-dotenv